perf(lifecycle): fix the five architectural review findings
1. Duplicate tier fetch: detached endpoint refresh and Pi-driven
   refreshModels share one in-flight live-catalog fetch per provider
   (fetchSharedCatalog). One request per tier per session_start instead
   of two; shared work takes no external signal (bounded by the
   fetchers' internal timeouts), callers keep abort semantics.
2. Jittered backoff: fetchWithRetry uses full-jitter exponential
   backoff (computeRetryBackoffMs, 10s cap) so concurrent refreshes
   stop retrying in lockstep after a gateway blip.
3. Splits: lib/util.ts -> lib/fetch.ts + lib/model-map.ts (untouched
   re-export shim; inline catalog mapper decomposed into named
   resolvers); auto-fallback settled decisions ->
   lib/auto-fallback/settled-decision.ts (pure, unit-tested), index.ts
   keeps orchestration; scope predicate deduplicated.
4. isFreeModel memoizes its pricing-exposure scan per catalog array
   identity (WeakMap); identical verdicts, no redundant O(n) rescans.

All behavior-preserving: 807 tests green (20 new), tsc, runtime
imports, lockfile checks pass.
