+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- **Lifecycle performance and complexity fixes (architectural review)** — five findings, all behavior-preserving with regression tests (807 tests green): (1) the detached built-in-tier endpoint refresh and Pi-driven `refreshModels` now share one in-flight live-catalog fetch per provider (`fetchSharedCatalog` in `lib/built-in-toggle.ts`), so every `session_start` costs one request per tier instead of two — shared work ignores individual abort signals (sockets stay bounded by the fetchers' internal timeouts) while callers keep their pre/post abort semantics; (2) `fetchWithRetry` backoff is now jittered exponential (`computeRetryBackoffMs`, full-jitter style, capped at 10s) so concurrent refreshes no longer retry in lockstep after a gateway blip; (3) `lib/util.ts` is split into `lib/fetch.ts` (network primitives) and `lib/model-map.ts` (mapping/filtering, with the 24-complexity inline catalog mapper decomposed into named per-field resolvers), re-exported via the untouched `lib/util.ts` shim so all 25+ importers keep working; auto-fallback's settled-run decisions (scope predicate, failure classification + abort refinement, auto-continue payload) moved to pure, unit-tested `lib/auto-fallback/settled-decision.ts` with `index.ts` keeping only orchestration; (4) `isFreeModel` memoizes its per-catalog pricing-exposure scan by array identity (`WeakMap`, no retention after catalog replacement), removing the redundant O(n) rescan per model with identical verdicts.
+
 ### Fixed
 
 - **OpenCode Go now syncs its live catalog, and a failed catalog refresh no longer surfaces as a Pi warning** — both halves of #504. `opencode-go` was the only built-in-toggle provider without `refreshEndpoint`, so its list was frozen at Pi's build-time snapshot plus Pi's own pi.dev mirror overlay (4h throttle): the public credential-free `GET https://opencode.ai/zen/go/v1/models` endpoint listed 35 models while pi-free showed 27, hiding `glm-5`, `grok-4.5`, `hy3-preview`, `kimi-k2.5`, `mimo-v2-omni`, `mimo-v2-pro`, `minimax-m2.5`, and `qwen3.5-plus` (point-in-time audit, 2026-09-06). The Go tier now runs the same detached post-session-start refresh as `opencode-free` — known IDs keep Pi's curated metadata and wire protocol, newly discovered IDs are synthesized through `resolveOpenCodeModelApi`/`applyOpenCodeProtocolDefaults` (which already special-cases Go's `minimax-*` Anthropic routing) — and it never blocks session start. Separately, Pi's `Could not refresh <provider>; showing cached models` warning was pi-free's own doing for `opencode-free`: the built-in-toggle `refreshModels` rethrew a failed catalog fetch, while every native provider retains the previous list and logs instead (`lib/native-provider.ts`). A transient stall — observed in `~/.pi/free.log` as `opencode-free: endpoint refresh failed {"error":"The operation was aborted due to timeout"}` once the 10s `DEFAULT_FETCH_TIMEOUT_MS` budget elapsed during a network blip — therefore became a user-visible warning even though the cached catalog stayed usable, which is exactly the reported "but opencode itself works". `refreshModels` now retains the cached view on a failed **or empty** fetch (poisoning guard, matching the native-provider recipe) and checks `context.signal.aborted` first so Pi's superseded-refresh aborts stay silent (convention 15); the detached endpoint refresh got the same empty-retain guard. The `opencode-go` half of that warning is Pi-side (its own pi.dev overlay fetch, which the provider composer awaits before ours) and cannot be caught from an extension.
