+++ agents.md
@@ -31,11 +31,12 @@ index.ts                          ← Extension entry point (piFreeEntry)
   ├─ lib/model-detection.ts       ← Model family grouping, name normalization
   ├─ lib/provider-cache.ts        ← Disk cache for fetched model lists
   ├─ lib/provider-compat.ts       ← DeepSeek proxy compat flag detection
-  ├─ lib/util.ts                  ← fetchWithRetry, model size parsing, OpenRouter mapping
+  ├─ lib/util.ts                  ← Back-compat re-export shim (new code: `lib/fetch.ts` for fetchWithRetry/timeout/deadline/retry-backoff, `lib/model-map.ts` for size parsing + OpenRouter/OpenAI-compatible mapping)
   ├─ lib/fallback-state.ts        ← Shared in-memory store (last HTTP status per model) used by both quota-monitor and auto-fallback
   ├─ lib/auto-fallback/           ← Auto-fallback to another free model on error (since 2.7.0; opt-in, default off)
   │   ├─ index.ts                 ← Event wiring (message_end observer / agent_settled single decision point / model_select)
   │   ├─ classifier.ts            ← HTTP status + errorMessage regex classification (mirrors pi-ai's `isRetryableAssistantError`)
+  │   ├─ settled-decision.ts      ← Pure settled-run decisions: scope predicate, failure classification + abort refinement, auto-continue payload (unit-tested; index.ts orchestrates)
   │   ├─ blacklist.ts             ← In-memory failure tracking with TTL + max-strikes
   │   ├─ selection.ts             ← CI-score candidate ordering + scope filter (provider / global / whitelist)
   │   ├─ notify.ts                ← 5-minute windowed notification aggregation
@@ -265,7 +266,7 @@ Debug logging writes to `~/.pi/free.log` under the `benchmark-lookup` namespace:
 9. **Model filtering is provider-specific** — native providers filter complete catalogs with `filterModels`; catalog fetchers may apply provider-specific modality or quality filters (NVIDIA retains its own 70B threshold)
 10. **All pi-free-registered catalogs use `enhanceWithCI()`** before registration to add CI scores
 11. **Legacy network-fetching extension providers are cache-first** (1h TTL via `lib/provider-cache.ts`); native providers, including FastRouter and StepFun, use Pi's models store + `refreshModels` instead. Pi's built-in OpenCode, OpenCode Go, and OpenRouter providers are not managed by either extension cache.
-12. **Startup and session-start work are observable and bounded** — legacy `loadCachedOrFetchModels` network attempts use `STARTUP_FETCH_DEADLINE_MS` (8s, override `PI_FREE_STARTUP_FETCH_TIMEOUT_MS`) via `withFetchDeadline` in `lib/util.ts`; failed and timed-out attempts are counted with duration and per-provider status. Native providers, including FastRouter, restore from Pi's store and their session-start refresh nudges and detached probes are measured without making intentionally background work block the event. `/free-startup` exposes the post-finalize handler/task timings.
+12. **Startup and session-start work are observable and bounded** — legacy `loadCachedOrFetchModels` network attempts use `STARTUP_FETCH_DEADLINE_MS` (8s, override `PI_FREE_STARTUP_FETCH_TIMEOUT_MS`) via `withFetchDeadline` in `lib/fetch.ts` (re-exported via `lib/util.ts`); failed and timed-out attempts are counted with duration and per-provider status. Native providers, including FastRouter, restore from Pi's store and their session-start refresh nudges and detached probes are measured without making intentionally background work block the event. `/free-startup` exposes the post-finalize handler/task timings.
 13. **Compiled packaging is shipped** — npm and Pi load `dist/index.js`, generated from the TypeScript source with peer dependencies externalized. The validation plan and packaging details live in [`docs/build-strategy.md`](docs/build-strategy.md).
 14. **Health output is credential-free** — `/pi-free-health` reports provider counts, timing/failure labels, and the configured log path; it must not print API keys, OAuth tokens, model payloads, or raw log contents.
 15. **Abort errors are expected, not failures** — Pi 0.84+ aborts a superseded model refresh; `AbortError`/`"This operation was aborted"` is cancellation, not a provider error. Every provider fetch `catch` block must guard `if (signal?.aborted)` and **return the function's empty error-path value** *before* logging — `[]` for array fetches, `{ all: [], free: [] }` for `{all,free}` catalogs, bare `return` for `Promise<void>` — so expected aborts never surface as visible `ERROR`/`notify` to the user. Check `signal?.aborted`, not `error.name === "AbortError"` (narrowed deliberately — an unrelated AbortError must still be logged). When fixing this class of bug (or any cross-cutting convention violation), **examine the whole repo and fix every instance**, not just the reported provider — providers share a near-identical fetch/`catch` shape, so a missing guard in one is a missing guard in all.
