+++ tests/util.test.ts
@@ -1,11 +1,13 @@
 import { describe, expect, it, vi } from "vitest";
 import {
 	cleanModelName,
+	computeRetryBackoffMs,
 	fetchWithRetry,
 	fetchWithTimeout,
 	isUsableModel,
 	logWarning,
 	mapOpenRouterModel,
+	MAX_RETRY_BACKOFF_MS,
 	parseModelResponse,
 	withFetchDeadline,
 } from "../lib/util.ts";
@@ -450,6 +452,43 @@ describe("Utility Functions", () => {
 			expect(result.ok).toBe(false);
 			expect(result.status).toBe(400);
 			expect(globalThis.fetch).toHaveBeenCalledTimes(1);
+		});
+	});
+
+	describe("computeRetryBackoffMs", () => {
+		it("scales exponentially with the attempt and stays under the window", () => {
+			const alwaysMax = () => 0.999999;
+			expect(computeRetryBackoffMs(0, 1000, { random: alwaysMax })).toBeLessThanOrEqual(1000);
+			expect(computeRetryBackoffMs(1, 1000, { random: alwaysMax })).toBeLessThanOrEqual(2000);
+			expect(computeRetryBackoffMs(2, 1000, { random: alwaysMax })).toBeLessThanOrEqual(4000);
+		});
+
+		it("caps the exponential growth at the configured maximum", () => {
+			const alwaysMax = () => 0.999999;
+			expect(
+				computeRetryBackoffMs(20, 1000, { random: alwaysMax }),
+			).toBeLessThanOrEqual(MAX_RETRY_BACKOFF_MS);
+			expect(computeRetryBackoffMs(20, 1000, { random: () => 1 })).toBe(
+				MAX_RETRY_BACKOFF_MS,
+			);
+			expect(
+				computeRetryBackoffMs(0, 1000, { capMs: 500, random: () => 1 }),
+			).toBe(500);
+		});
+
+		it("decorrelates concurrent retries via jitter", () => {
+			const samples = new Set(
+				Array.from({ length: 20 }, () => computeRetryBackoffMs(2, 1000)),
+			);
+			// 20 full-jitter draws colliding would be a ~0-probability event.
+			expect(samples.size).toBeGreaterThan(1);
+		});
+
+		it("treats non-positive inputs as no backoff", () => {
+			const alwaysMax = () => 0.999999;
+			expect(computeRetryBackoffMs(-1, 1000, { random: alwaysMax })).toBeLessThanOrEqual(1000);
+			expect(computeRetryBackoffMs(0, 0, { random: alwaysMax })).toBe(0);
+			expect(computeRetryBackoffMs(0, -50, { random: alwaysMax })).toBe(0);
 		});
 	});
